Mario-AI-Framework/requirements.txt
absl-py==1.2.0
ale-py==0.7.4
apturl==0.5.2
asttokens==2.0.8
astunparse==1.6.3
asyncio==3.4.3
atari-py==0.2.6
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.14.20
botocore==1.17.63
Box2D==2.3.2
box2d-py==2.3.5
Brlapi==0.8.3
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.1
chai-sacred==0.8.3
chardet==4.0.0
click==8.0.3
cloudpickle==2.2.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.5
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.32
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
docopt==0.6.2
docutils==0.15.2
duplicity==0.8.21
entrypoints==0.4
executing==1.0.0
fasteners==0.14.1
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.37.1
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.5
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
gym==0.21.0
gym-notices==0.0.8
h5py==3.7.0
highway-env==1.6
httplib2==0.20.2
huggingface-hub==0.8.1
huggingface-sb3==2.2.3
idna==3.3
imageio==2.21.3
-e git+http://github.com/HumanCompatibleAI/imitation@2fea7f9734c07a80f284706328e5c49878fcccea#egg=imitation
importlib-metadata==4.13.0
importlib-resources==5.9.0
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
jeepney==0.7.1
jmespath==0.10.0
joblib==1.1.0
JPype1==1.4.0
jsonpickle==2.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
lockfile==0.12.2
louis==3.20.0
lz4==4.0.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
munch==2.5.0
nest-asyncio==1.5.5
netifaces==0.11.0
numpy==1.23.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pygame==2.1.0
pyglet==1.5.0
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.14.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
pyzmq==23.2.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.3.7
scikit-learn==1.1.2
scipy==1.9.1
seals==0.1.4
SecretStorage==3.3.1
shortuuid==1.0.1
six==1.16.0
smmap==5.0.0
stable-baselines==2.10.2
stable-baselines3 @ git+https://github.com/DLR-RM/stable-baselines3@537a82a7fdadc710d1ca0e98dddc967e3870f19d
stack-data==0.5.0
systemd-python==234
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
torch==1.12.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.25.11
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wasabi==0.10.1
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
